+++ packages/Chembl/src/package-test.ts
@@ -4,6 +4,7 @@ import {runTests, tests, TestContext, initAutoTests as initTests } from '@datagr
 import './tests/cartridge';
 import './tests/converters';
 import './tests/frac';
+import './tests/atc';
 
 export const _package = new DG.Package();
 export {tests};